Omit the redundant "R" at the start of title.
Provide a link to the used webservices (IFN) to the description
field of your DESCRIPTION file in the form
<http:...> or <https:...>
with angle brackets for auto-linking and no space after 'http:' and
'https:'.

Explain all acronyms in the description text. -> IFN

Write TRUE and FALSE instead of T and F. Don't use "T" or
"F" as vector names.
  'T' and 'F' instead of TRUE and FALSE:
   man/get_ifn_all.Rd:
     get_ifn_all(unlist = T)
     get_ifn_all(unlist = T)
   man/get_url.Rd:
     get_url(all = F)